Fix InvalidCastException on startup on non-English regional settings
Exception `System.InvalidCastException` with a message of
`Conversion from string "8.2" to type 'Decimal' is not valid.` was
raised when running the application on non-English regional settings
(when character different than . (dot) is a numerical separator).

Explicit usage of InvariantCulture when converting string to decimal
is not based on the regional separator, so it's always working.